Combine the phosphorylation and ligation steps.
This is inspired by NEB's KLD mix, which it sells as part of its site
directed mutagenesis kit.  I assume the enzymes are the same, because
they're all from NEB, and I assume the 1:1:1 proportions are close
enough, because I usually use 1 μL of each enzyme separately.  I also
assume that the T4 ligase buffer is appropriate for all the enzymes,
because I already know it works with PNK and DpnI isn't known to be that
sensitive to the buffer.